requirements.txt
annotated-types==0.6.0
asttokens==2.4.1
bcrypt==4.1.1
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
decorator==5.1.1
executing==2.0.1
Flask
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy
Flask-WTF==0.14.2
idna==3.6
ipython==8.18.1
itsdangerous==2.0.1
jedi==0.19.1
Jinja2==3.0.3
langcodes==3.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
murmurhash==1.0.10
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
preshed==3.0.9
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2

scipy
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
smart-open==6.4.0
SQLAlchemy==2.0.23
srsly==2.4.8
stack-data==0.6.3
thinc==8.2.1
tqdm==4.66.1
traitlets==5.14.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wasabi==1.1.2
wcwidth==0.2.12
weasel==0.3.4
Werkzeug==0.16.1
WTForms==2.2.1
openai
gunicorn
python-dotenv